Enabled Wextra compilation flag

+++ tests/pmap.c
@@ -23,9 +23,9 @@ static int test_kernel_pmap() {
     tlb_print();
 
     int *x = (int *)vaddr1;
-    for (int i = 0; i < size / sizeof(int); i++)
+    for (int i = 0; i < (int)(size / sizeof(int)); i++)
       *(x + i) = i;
-    for (int i = 0; i < size / sizeof(int); i++)
+    for (int i = 0; i < (int)(size / sizeof(int)); i++)
       assert(*(x + i) == i);
 
     klog("TLB after:");
